Fixed reminders perf tests
