+++ tests/apps/perf/actorfeatures/Dockerfile
@@ -1,25 +1,19 @@
-#
-# Copyright 2021 The Dapr Authors
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#     http://www.apache.org/licenses/LICENSE-2.0
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-FROM golang:1.20 as build_env
+FROM golang:1.20-bullseye as build_env
 
 ENV CGO_ENABLED=0
 WORKDIR /app
-COPY . .
-RUN go get -d -v
-RUN go build -o app .
+COPY *.go go.mod go.sum ./
+RUN go get -d -v && go build -o tester .
 
-FROM gcr.io/distroless/static:nonroot
+FROM golang:1.20-bullseye as fortio_build_env
+
+WORKDIR /fortio
+ADD "https://api.github.com/repos/dapr/fortio/branches/v1.38.4-dapr" skipcache
+RUN git clone https://github.com/dapr/fortio.git
+RUN cd fortio && git checkout v1.38.4-dapr && go build
+
+FROM debian:bullseye-slim
 WORKDIR /
-COPY --from=build_env /app/app /
-CMD ["/app"]
+COPY --from=build_env /app/tester /
+COPY --from=fortio_build_env /fortio/fortio/fortio /usr/local/bin
+CMD ["/tester"]
