+++ tests/apps/perf/actorfeatures/go.mod
@@ -1,5 +1,5 @@
-module github.com/dapr/dapr/tests/apps/perf/actorfeatures
+module github.com/dapr/dapr/tests/apps/perf/tester
 
 go 1.20
 
-require github.com/gorilla/mux v1.7.3
+require github.com/go-chi/chi/v5 v5.0.10
