+++ tests/apps/perf/tester/app.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2021 The Dapr Authors
+Copyright 2023 The Dapr Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -14,120 +14,16 @@ limitations under the License.
 package main
 
 import (
-	"encoding/json"
-	"fmt"
-	"io"
 	"log"
 	"net/http"
-	"os"
-	"os/exec"
-	"strconv"
 )
-
-// TODO: change to take from "github.com/dapr/dapr/tests/perf" once in repository. otherwise fails on go get step in Dockerfile.
-type TestParameters struct {
-	QPS               int    `json:"qps"`
-	ClientConnections int    `json:"clientConnections"`
-	TargetEndpoint    string `json:"targetEndpoint"`
-	TestDuration      string `json:"testDuration"`
-	PayloadSizeKB     int    `json:"payloadSizeKB"`
-	Payload           string `json:"payload"`
-	StdClient         bool   `json:"stdClient"`
-	Grpc              bool   `json:"grpc"`
-	Dapr              string `json:"dapr"`
-}
 
 func handler(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(200)
 }
 
-func testHandler(w http.ResponseWriter, r *http.Request) {
-	fmt.Println("test execution request received")
-
-	var testParams TestParameters
-	b, err := io.ReadAll(r.Body)
-	if err != nil {
-		w.WriteHeader(500)
-		w.Write([]byte(fmt.Sprintf("error reading request body: %s", err)))
-		return
-	}
-
-	err = json.Unmarshal(b, &testParams)
-	if err != nil {
-		w.WriteHeader(400)
-		w.Write([]byte(fmt.Sprintf("error parsing test params: %s", err)))
-		return
-	}
-
-	fmt.Println("executing test")
-	results, err := runTest(testParams)
-	if err != nil {
-		w.WriteHeader(500)
-		w.Write([]byte(fmt.Sprintf("error encountered while running test: %s", err)))
-		return
-	}
-
-	fmt.Println("test finished")
-	w.Header().Add("Content-Type", "application/json")
-	w.Write(results)
-}
-
 func main() {
 	http.HandleFunc("/", handler)
-	http.HandleFunc("/test", testHandler)
+	http.HandleFunc("/test", fortioTestHandler)
 	log.Fatal(http.ListenAndServe(":3001", nil))
-}
-
-// runTest accepts a set of test parameters, runs Fortio with the configured setting and returns
-// the test results in json format.
-func runTest(params TestParameters) ([]byte, error) {
-	args := buildFortioArgs(params)
-	fmt.Printf("running test with params: %s", args)
-
-	cmd := exec.Command("fortio", args...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err := cmd.Run()
-	if err != nil {
-		return nil, err
-	}
-	return os.ReadFile("result.json")
-}
-
-func buildFortioArgs(params TestParameters) []string {
-	var args []string
-
-	if len(params.Payload) > 0 {
-		args = []string{
-			"load",
-			"-json", "result.json",
-			"-content-type", "application/json",
-			"-qps", strconv.Itoa(params.QPS),
-			"-c", strconv.Itoa(params.ClientConnections),
-			"-t", params.TestDuration,
-			"-payload", params.Payload,
-		}
-	} else {
-		args = []string{
-			"load",
-			"-json", "result.json",
-			"-qps", strconv.Itoa(params.QPS),
-			"-c", strconv.Itoa(params.ClientConnections),
-			"-t", params.TestDuration,
-			"-payload-size", strconv.Itoa(params.PayloadSizeKB),
-		}
-	}
-	if params.StdClient {
-		args = append(args, "-stdclient")
-	}
-
-	if params.Grpc {
-		args = append(args, "-grpc")
-	}
-	if params.Dapr != "" {
-		args = append(args, "-dapr", params.Dapr)
-	}
-
-	args = append(args, params.TargetEndpoint)
-	return args
 }
