+++ tests/perf/actor_reminder/actor_reminder_test.go
@@ -34,10 +34,9 @@ import (
 )
 
 const (
-	numHealthChecks        = 60 // Number of times to check for endpoint health per app.
-	actorType              = "PerfTestActorReminder"
-	serviceApplicationName = "perf-actor-reminder-service"
-	clientApplicationName  = "perf-actor-reminder-client"
+	numHealthChecks = 60 // Number of times to check for endpoint health per app.
+	actorType       = "PerfTestActorReminder"
+	appName         = "perf-actor-reminder-service"
 
 	// Target for the QPS
 	targetQPS = 50
@@ -50,7 +49,7 @@ func TestMain(m *testing.M) {
 
 	testApps := []kube.AppDescription{
 		{
-			AppName:           serviceApplicationName,
+			AppName:           appName,
 			DaprEnabled:       true,
 			ImageName:         "perf-actorfeatures",
 			Replicas:          1,
@@ -66,31 +65,6 @@ func TestMain(m *testing.M) {
 			AppMemoryRequest:  "2500Mi",
 			AppEnv: map[string]string{
 				"TEST_APP_ACTOR_TYPE": actorType,
-			},
-			Labels: map[string]string{
-				"daprtest": serviceApplicationName,
-			},
-		},
-		{
-			AppName:           clientApplicationName,
-			DaprEnabled:       true,
-			ImageName:         "perf-tester",
-			Replicas:          1,
-			IngressEnabled:    true,
-			AppPort:           3001,
-			DaprCPULimit:      "4.0",
-			DaprCPURequest:    "0.1",
-			DaprMemoryLimit:   "512Mi",
-			DaprMemoryRequest: "250Mi",
-			AppCPULimit:       "4.0",
-			AppCPURequest:     "0.1",
-			AppMemoryLimit:    "800Mi",
-			AppMemoryRequest:  "2500Mi",
-			Labels: map[string]string{
-				"daprtest": clientApplicationName,
-			},
-			PodAffinityLabels: map[string]string{
-				"daprtest": serviceApplicationName,
 			},
 		},
 	}
@@ -108,21 +82,12 @@ func TestActorReminderRegistrationPerformance(t *testing.T) {
 	)
 
 	// Get the ingress external url of test app
-	testAppURL := tr.Platform.AcquireAppExternalURL(serviceApplicationName)
+	testAppURL := tr.Platform.AcquireAppExternalURL(appName)
 	require.NotEmpty(t, testAppURL, "test app external URL must not be empty")
 
 	// Check if test app endpoint is available
 	t.Logf("test app url: %s", testAppURL+"/health")
 	_, err := utils.HTTPGetNTimes(testAppURL+"/health", numHealthChecks)
-	require.NoError(t, err)
-
-	// Get the ingress external url of tester app
-	testerAppURL := tr.Platform.AcquireAppExternalURL(clientApplicationName)
-	require.NotEmpty(t, testerAppURL, "tester app external URL must not be empty")
-
-	// Check if tester app endpoint is available
-	t.Logf("tester app url: %s", testerAppURL)
-	_, err = utils.HTTPGetNTimes(testerAppURL, numHealthChecks)
 	require.NoError(t, err)
 
 	// Perform dapr test
@@ -133,7 +98,7 @@ func TestActorReminderRegistrationPerformance(t *testing.T) {
 	require.NoError(t, err)
 
 	t.Logf("running dapr test with params: %s", body)
-	daprResp, err := utils.HTTPPost(fmt.Sprintf("%s/test", testerAppURL), body)
+	daprResp, err := utils.HTTPPost(fmt.Sprintf("%s/test", testAppURL), body)
 	t.Log("checking err...")
 	require.NoError(t, err)
 	require.NotEmpty(t, daprResp)
@@ -143,13 +108,13 @@ func TestActorReminderRegistrationPerformance(t *testing.T) {
 	// Let test run for 90s triggering the timers and collect metrics.
 	time.Sleep(90 * time.Second)
 
-	appUsage, err := tr.Platform.GetAppUsage(serviceApplicationName)
+	appUsage, err := tr.Platform.GetAppUsage(appName)
 	require.NoError(t, err)
 
-	sidecarUsage, err := tr.Platform.GetSidecarUsage(serviceApplicationName)
+	sidecarUsage, err := tr.Platform.GetSidecarUsage(appName)
 	require.NoError(t, err)
 
-	restarts, err := tr.Platform.GetTotalRestarts(serviceApplicationName)
+	restarts, err := tr.Platform.GetTotalRestarts(appName)
 	require.NoError(t, err)
 
 	t.Logf("dapr test results: %s", string(daprResp))
@@ -159,7 +124,7 @@ func TestActorReminderRegistrationPerformance(t *testing.T) {
 
 	var daprResult perf.TestResult
 	err = json.Unmarshal(daprResp, &daprResult)
-	require.NoError(t, err)
+	require.NoErrorf(t, err, "Failed to unmarshal: %s", string(daprResp))
 
 	percentiles := map[int]string{2: "90th", 3: "99th"}
 
@@ -182,8 +147,8 @@ func TestActorReminderRegistrationPerformance(t *testing.T) {
 	}
 
 	summary.ForTest(t).
-		Service(serviceApplicationName).
-		Client(clientApplicationName).
+		Service(appName).
+		Client(appName).
 		CPU(appUsage.CPUm).
 		Memory(appUsage.MemoryMb).
 		SidecarCPU(sidecarUsage.CPUm).
